+++ packages/client/src/InfluxDBClient.ts
@@ -120,7 +120,11 @@ export default class InfluxDBClient {
       this._options.host = host.substring(0, host.length - 1)
     if (typeof this._options.token !== 'string')
       throw new IllegalArgumentError('No token specified!')
-    this._queryApi = new QueryApiImpl({...this._options, queryTimeout: this._options.queryOptions?.timeout ?? this._options?.queryTimeout})
+    this._queryApi = new QueryApiImpl({
+      ...this._options,
+      queryTimeout:
+        this._options.queryOptions?.timeout ?? this._options?.queryTimeout,
+    })
 
     this._transport = impl.writeTransport(this._options)
     this._writeApi = new WriteApiImpl({